+++ internal/core/audit.go
@@ -17,6 +17,7 @@ const (
 	OperationGenerateX25519     = "generate_x25519"
 	OperationX25519SharedSecret = "x25519_shared_secret"
 	OperationDeriveEd25519      = "derive_ed25519_private_key"
+	OperationGenerateCert       = "generate_cert"
 	OperationDeriveKey          = "derive_key"
 	OperationRotateKEK          = "rotate_kek"
 )